Split generic behavior import diagnostics tests
